# src/lookup/src/services/swagger.json
{
  "x-generator": "NSwag v13.15.10.0 (NJsonSchema v10.6.10.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "Lookup API",
    "description": "Backend API for Lookup",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:7256"
    }
  ],
  "paths": {
    "/api/v1/lookup/{id}": {
      "get": {
        "tags": [
          "Lookup"
        ],
        "operationId": "Lookup_GetLookup",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "reply",
          "content": {
            "application/json": {
              "schema": {
                "type": "boolean"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupMessage"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/lookup": {
      "post": {
        "tags": [
          "Lookup"
        ],
        "operationId": "Lookup_PostLookup",
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLookupModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/lookup/edit/{id}": {
      "post": {
        "tags": [
          "Lookup"
        ],
        "operationId": "Lookup_EditLookup",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditLookupModel"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/lookup/{userId}/messages": {
      "get": {
        "tags": [
          "Lookup"
        ],
        "operationId": "Lookup_GetMessagesFromUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupMessage"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/lookup/{lookupId}/thread": {
      "get": {
        "tags": [
          "Lookup"
        ],
        "operationId": "Lookup_GetMessageThread",
        "parameters": [
          {
            "name": "lookupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupMessage"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/user/{userId}/followers": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "User_GetFollowersList",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/user/userid:guid/following": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "User_GetFollowingList",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/user/{userId}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "User_GetUserInfo",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/user/me/follow/{userToFollowId}": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "User_FollowUser",
        "parameters": [
          {
            "name": "userToFollowId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/user/me/messages": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "User_GetReceivedMessages",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupMessage"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/user/me": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "User_GetMeInfo",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/weather": {
      "get": {
        "tags": [
          "Weather"
        ],
        "operationId": "Weather_Get",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Weather"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "LookupMessage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "content": {
            "type": "string"
          },
          "publisherUserId": {
            "type": "string",
            "format": "guid"
          },
          "publisherUsername": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "likes": {
            "type": "integer",
            "format": "int32"
          },
          "edited": {
            "type": "boolean"
          },
          "replyId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "editedTimetamp": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "CreateLookupModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "content": {
            "type": "string"
          },
          "edited": {
            "type": "boolean"
          },
          "replyId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          }
        }
      },
      "EditLookupModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "username": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "headerUrl": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "joinedDate": {
            "type": "string",
            "format": "date-time"
          },
          "followersCount": {
            "type": "integer",
            "format": "int32"
          },
          "followingCount": {
            "type": "integer",
            "format": "int32"
          },
          "lookupsCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Weather": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "temp": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "oauth2",
        "description": "Type into the text box: Bearer {your JWT token}.",
        "name": "Authorization",
        "scheme": "Bearer",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://danovas.b2clogin.com/danovas.onmicrosoft.com/B2C_1_signupsignin1/oauth2/v2.0/authorize",
            "tokenUrl": "https://danovas.b2clogin.com/danovas.onmicrosoft.com/B2C_1_signupsignin1/oauth2/v2.0/token",
            "scopes": {
              "https://danovas.onmicrosoft.com/89be5e10-1770-45d7-813a-d47242ae2163/API.Access": "Access the api as the signed-in user"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}